Collegato visivamente Fill con volume di liquido nel becher

+++ Assets/Objects/Scripts/Fillable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface Fillable
 {
-    public int Fill(int volume);
+    public float Fill(float volume);
 }
